# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator project.",
        "contact": {
            "email": "manish011994@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/user/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "description": "Retrieve the preferences of the authenticated user.",
                "operationId": "b9e662271a19e3c25c2b6bcdf8ff231d",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preference found."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserPreferenceResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Preference not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preference not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Store user preferences",
                "description": "Store the user's preferences for sources, categories, and authors.",
                "operationId": "0e809239a79e066ddba5ac221e4af2e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sources": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "source1",
                                            "source2"
                                        ]
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "category1",
                                            "category2"
                                        ]
                                    },
                                    "authors": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "author1",
                                            "author2"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preference(s) updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserPreferenceResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or missing preferences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "At least one source, authors or categories are required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/personalized-feed": {
            "get": {
                "tags": [
                    "User Feed"
                ],
                "summary": "Get personalized feed for the user",
                "description": "Retrieve a personalized feed of articles based on the user's preferences.",
                "operationId": "abd5b050e57ed6bf91ffdc4ad9342d5c",
                "responses": {
                    "200": {
                        "description": "Personalized feed retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Personalized feed found."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ArticleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No personalized feeds found or preference not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No feeds found for your preferences, please adjust your preferences."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ArticleResource": {
                "title": "Article",
                "description": "Schema representing an article",
                "properties": {
                    "id": {
                        "description": "ID of the article",
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "description": "Author of the article",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "title": {
                        "description": "Title of the article",
                        "type": "string",
                        "example": "Understanding OpenAPI"
                    },
                    "source": {
                        "description": "Source of the article",
                        "type": "string",
                        "example": "Tech News",
                        "nullable": true
                    },
                    "source_url": {
                        "description": "Source URL of the article",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/article-source",
                        "nullable": true
                    },
                    "description": {
                        "description": "Description of the article",
                        "type": "string",
                        "example": "This article explains the basics of OpenAPI."
                    },
                    "keywords": {
                        "description": "Keywords associated with the article",
                        "type": "string",
                        "example": "OpenAPI, Swagger, API Documentation"
                    },
                    "category": {
                        "description": "Category of the article",
                        "type": "string",
                        "example": "Technology"
                    },
                    "image_url": {
                        "description": "Image URL associated with the article",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "content": {
                        "description": "Full content of the article",
                        "type": "string",
                        "example": "This is the content of the article..."
                    },
                    "published_at": {
                        "description": "Published date and time of the article",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-11T14:30:00Z"
                    },
                    "status": {
                        "description": "Status of the article",
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "published"
                    },
                    "created_at": {
                        "description": "Date and time when the article was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-10T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Date and time when the article was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-11T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "UserPreferenceResource": {
                "title": "User Preference",
                "description": "Schema representing a user's preferences",
                "properties": {
                    "id": {
                        "description": "ID of the user preference",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the associated user",
                        "type": "integer",
                        "example": 42
                    },
                    "sources": {
                        "description": "Preferred sources of the user",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "source_1"
                        }
                    },
                    "categories": {
                        "description": "Preferred categories of the user",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "category_1"
                        }
                    },
                    "authors": {
                        "description": "Preferred authors of the user",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "author_1"
                        }
                    },
                    "created_at": {
                        "description": "Timestamp when the preference was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-10T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the preference was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-11T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}